Added int cast to playerSlot to DebugLogFileHandler method call
- Battle couldn't be started and there was an error since it didn't recognize the enum

+++ Assets/Altzone/Scripts/Photon/LobbyManager.cs
@@ -1010,7 +1010,7 @@ namespace Altzone.Scripts.Lobby
             yield return new WaitUntil(()=>SceneManager.GetActiveScene().name == _quantumBattleMap.Scene);
 
             DebugLogFileHandler.ContextEnter(DebugLogFileHandler.ContextID.Battle);
-            DebugLogFileHandler.FileOpen(battleID, playerSlot);
+            DebugLogFileHandler.FileOpen(battleID, (int)playerSlot);
 
             Task<bool> task = StartRunner(sessionRunnerArguments);
